requirements.txt
PyQt5
ultralytics
torch
torchvision
opencv-python
pydicom
numpy
matplotlib
PyYAML
